remove objects.rs
- `objects.rs` was a mix of unrelated items, move them to separate
  modules for easier maintainability
