+++ src/output/mod.rs
@@ -5,23 +5,29 @@
 // https://opensource.org/licenses/MIT.
 
 mod config;
+mod dut;
 mod emitters;
+mod error;
+mod log;
 mod macros;
-mod measurement_series;
+mod measurement;
 mod models;
-mod objects;
 mod run;
 mod state;
 mod step;
 
 pub use config::*;
+pub use dut::*;
 pub use emitters::*;
+pub use error::*;
+pub use log::*;
+pub use measurement::*;
 pub use models::LogSeverity;
 pub use models::TestResult;
 pub use models::TestStatus;
 pub use models::ValidatorType;
 pub use models::SPEC_VERSION;
-pub use objects::*;
 pub use run::*;
-pub use serde_json::Value;
 pub use step::*;
+
+pub use serde_json::Value;
